requirements.txt
Django==3.2.5
djangorestframework==3.12.4
mysqlclient==2.0.3
django-redis==5.0.0
django-celery-results==2.2.0
django-celery-beat==2.2.1
eventlet==0.31.1

